ai-worker/requirements.txt
# AI Worker - Face Recognition & Metadata Extraction

face_recognition==1.3.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pgvector>=0.2.0
python-dotenv>=1.0.0
boto3>=1.34.0
numpy>=1.24.0
